Basic networking impl
TODOs for networking:
need to add serverside chunk generation
need to add better negotiation
need to add chunk send requests and chunk data transfers
need to fix remote animations not updating properly
need to figure out how to do reconciliation for local player

+++ client/src/item/item.h
@@ -2,7 +2,7 @@
 
 #include <SFML/Graphics.hpp>
 
-enum class ItemType: uint8_t {
+enum class ItemType: uint16_t {
     WoodSword = 0,
     CopperSword = 1,
     IronSword = 2,
